requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
PyJWT==2.8.0
python-multipart==0.0.6
websockets==12.0

